Update generated contract and model files; refactor Metagame and scoring logic
- Regenerated contract files for evolute_duel-game, evolute_duel-metagame, evolute_duel-player_profile_actions, evolute_duel-rewards_manager, evolute_duel-tutorial, and various models to reflect recent changes.
- Refactored Metagame methods to improve clarity and logging, including detailed debug statements for tile placement validation.
- Updated scoring logic to ensure accurate calculations during tile placements.
- Changed data types in scoring helper functions for better compatibility.
- Enhanced validation checks with comprehensive logging for easier debugging.

+++ src/systems/helpers/scoring.cairo
@@ -146,7 +146,7 @@ pub fn connect_adjacent_edges(
             continue;
         }
 
-        let offset: i32 = *direction_offsets[side.into()];
+        let offset: i64 = *direction_offsets[side.into()];
         let node_position: u32 = tile_position * 4 + side.into();
         let adjacent_node_position: u32 = (node_position.try_into().unwrap() + offset).try_into().unwrap();
         let (col_offset, row_offset) = *col_row_offsets.at(side.into());
